rename to dash_slicer and VolumeSlicer
