+++ examples/slicer_with_1_view.py
@@ -4,17 +4,17 @@ A truly minimal example.
 
 import dash
 import dash_html_components as html
-from dash_3d_viewer import DashVolumeSlicer
+from dash_slicer import VolumeSlicer
 import imageio
 
 
 app = dash.Dash(__name__)
 
 vol = imageio.volread("imageio:stent.npz")
-slicer = DashVolumeSlicer(app, vol)
+slicer = VolumeSlicer(app, vol)
 
 app.layout = html.Div([slicer.graph, slicer.slider, *slicer.stores])
 
 
 if __name__ == "__main__":
-    app.run_server(debug=False)
+    app.run_server(debug=True)
