+++ examples/slicer_with_2_views.py
@@ -4,15 +4,15 @@ An example with two slicers on the same volume.
 
 import dash
 import dash_html_components as html
-from dash_3d_viewer import DashVolumeSlicer
+from dash_slicer import VolumeSlicer
 import imageio
 
 
 app = dash.Dash(__name__)
 
 vol = imageio.volread("imageio:stent.npz")
-slicer1 = DashVolumeSlicer(app, vol, axis=1)
-slicer2 = DashVolumeSlicer(app, vol, axis=2)
+slicer1 = VolumeSlicer(app, vol, axis=1)
+slicer2 = VolumeSlicer(app, vol, axis=2)
 
 app.layout = html.Div(
     style={
