+++ examples/slicer_with_3_views.py
@@ -7,7 +7,7 @@ import plotly.graph_objects as go
 import dash
 import dash_html_components as html
 import dash_core_components as dcc
-from dash_3d_viewer import DashVolumeSlicer
+from dash_slicer import VolumeSlicer
 from skimage.measure import marching_cubes
 import imageio
 
@@ -15,9 +15,9 @@ app = dash.Dash(__name__)
 
 # Read volumes and create slicer objects
 vol = imageio.volread("imageio:stent.npz")
-slicer1 = DashVolumeSlicer(app, vol, reverse_y=False, axis=0)
-slicer2 = DashVolumeSlicer(app, vol, reverse_y=False, axis=1)
-slicer3 = DashVolumeSlicer(app, vol, reverse_y=False, axis=2)
+slicer1 = VolumeSlicer(app, vol, reverse_y=False, axis=0)
+slicer2 = VolumeSlicer(app, vol, reverse_y=False, axis=1)
+slicer3 = VolumeSlicer(app, vol, reverse_y=False, axis=2)
 
 # Calculate isosurface and create a figure with a mesh object
 verts, faces, _, _ = marching_cubes(vol, 300, step_size=2)
@@ -68,4 +68,4 @@ app.layout = html.Div(
 
 
 if __name__ == "__main__":
-    app.run_server(debug=False)
+    app.run_server(debug=True)
